splitting out crawl
