--- cron/cron_test.go
@@ -1,13 +1,0 @@
-package cron
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-)
-
-func TestDebug(t *testing.T) {
-	r := httptest.NewRequest(http.MethodGet, "/cron/debug", nil)
-	w := httptest.NewRecorder()
-	debug(w, r)
-}
